+++ lib/screen/auth/sign_up_main.dart
@@ -1,20 +1,16 @@
 import 'package:flutter/material.dart';
-import 'package:mukgen_flutter_v1/screen/SignUp/sign_up_id_pw_page.dart';
+import 'package:mukgen_flutter_v1/screen/auth/sign_up_nickname.dart';
 import 'package:transition/transition.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class SignUpNamePage extends StatefulWidget {
-  const SignUpNamePage({Key? key}) : super(key: key);
+class SignupStartPage extends StatefulWidget {
+  const SignupStartPage({Key? key}) : super(key: key);
 
   @override
-  State<SignUpNamePage> createState() => _SignupNamePageState();
+  State<SignupStartPage> createState() => _SignupStartPageState();
 }
 
-class _SignupNamePageState extends State<SignUpNamePage> {
-  String _inputValue = '';
-
-  bool get isFormValid => _inputValue.isNotEmpty;
-
+class _SignupStartPageState extends State<SignupStartPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -42,9 +38,9 @@ class _SignupNamePageState extends State<SignUpNamePage> {
           SizedBox(height: 40.0.h),
           Container(
             alignment: Alignment.topLeft,
-            margin: EdgeInsets.only(left: 20.0.w),
+            padding: EdgeInsets.only(left: 20.0.w),
             child: Text(
-              '별명을 입력해주세요.',
+              '시작하기',
               style: TextStyle(
                   fontSize: 24,
                   fontWeight: FontWeight.w600,
@@ -53,33 +49,25 @@ class _SignupNamePageState extends State<SignUpNamePage> {
           ),
           SizedBox(height: 24.0.h),
           Container(
-            width: 352.0.w,
-            child: TextFormField(
-              onChanged: (value) {
-                setState(() {
-                  _inputValue = value;
-                });
-              },
+            alignment: Alignment.topLeft,
+            margin: EdgeInsets.only(left: 20.0.w),
+            child: Text(
+              '먹젠과 함께 학교에서 즐거운 식생활을 누려보세요.',
               style: TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 20,
+                fontSize: 16,
+                fontFamily: 'MukgenRegular',
+                fontWeight: FontWeight.w400,
               ),
-              decoration: InputDecoration(
-                hintText: '별명',
-                hintStyle: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w600,
-                    fontFamily: 'MukgenSemiBold',
-                    color: Color(0xFFC9CAD0)),
-                enabledBorder: _inputValue.isEmpty
-                    ? UnderlineInputBorder(
-                        borderSide:
-                            BorderSide(color: Color(0xFFC9CAD0), width: 2))
-                    : UnderlineInputBorder(
-                        borderSide:
-                            BorderSide(color: Color(0xFF000000), width: 2)),
-                focusedBorder: UnderlineInputBorder(
-                    borderSide: BorderSide(color: Color(0xFFFE6B01), width: 2)),
+            ),
+          ),
+          Container(
+            alignment: Alignment.center,
+            child: Transform.translate(
+              offset: Offset(0, 120),
+              child: Image(
+                image: AssetImage('assets/images/signupmukgen.png'),
+                width: 160.0.w,
+                height: 160.0.h,
               ),
             ),
           ),
@@ -90,19 +78,17 @@ class _SignupNamePageState extends State<SignUpNamePage> {
             child: TextButton(
               onPressed: () {
                 setState(() {
-                  if (isFormValid == true) {
-                    Navigator.push(
-                      context,
-                      Transition(
-                        child: SignUpIdPwPage(),
-                        transitionEffect: TransitionEffect.RIGHT_TO_LEFT,
-                      ),
-                    );
-                  }
+                  Navigator.push(
+                    context,
+                    Transition(
+                      child: SignUpNamePage(),
+                      transitionEffect: TransitionEffect.RIGHT_TO_LEFT,
+                    ),
+                  );
                 });
               },
               child: Text(
-                '다음',
+                '시작하기',
                 style: TextStyle(
                   fontWeight: FontWeight.w600,
                   fontSize: 16,
@@ -111,9 +97,7 @@ class _SignupNamePageState extends State<SignUpNamePage> {
                 ),
               ),
               style: ButtonStyle(
-                backgroundColor: isFormValid == false
-                    ? MaterialStateProperty.all(Color(0xFFC9CAD0))
-                    : MaterialStateProperty.all(Color(0xFF6B6C71)),
+                backgroundColor: MaterialStateProperty.all(Color(0xFF6B6C71)),
                 shape: MaterialStateProperty.all<RoundedRectangleBorder>(
                   RoundedRectangleBorder(
                     borderRadius: BorderRadius.circular(10.0),
