+++ lib/screen/auth/sign_up_number.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mukgen_flutter_v1/screen/SignUp/signup_nickname.dart';
+import 'package:mukgen_flutter_v1/screen/auth/sign_up_nickname.dart';
 import 'package:mukgen_flutter_v1/screen/starting_page.dart';
 import 'package:transition/transition.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
