+++ lib/screen/starting_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mukgen_flutter_v1/common/common.dart';
-import 'package:mukgen_flutter_v1/screen/SignUp/signup_main.dart';
+import 'package:mukgen_flutter_v1/screen/auth/sign_up_main.dart';
 import 'package:transition/transition.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:mukgen_flutter_v1/screen/auth/login_page.dart';
